Specifically choose iOS 11.2.
The `cordova-paramedic` npm tries to obtain the logs from the simulator
run, however the `cordova` npm is only aware of minor versions in iOS
releases, not patch versions (e.g 11.0.1, which it classifies as 11.0).

This specifically chooses a version, though it may be best if we fix
this in a future version fo `cordova-paramedic`.